Add: 'CreateLadderGame.java' interface and 'RandomLadderGame.java'
