+++ src/main/java/ladderGame/CreateLadderGame.java
@@ -1,0 +1,5 @@
+package ladderGame;
+
+public interface CreateLadderGame {
+    LadderGame create(LadderGame ladderGame);
+}
